fix(learn): wrap table rows in <tbody>
Firefox implicitly inserts \<tbody\> element to the table displayed on the page, while editable html doesn't include it.

In the task you need to select every other row of the table without modifying html. One way to do that is to write following selector:
```
table > th:nth-child(2n + 1) {
}
```
But because \<tr\> is wrapped in \<tbody\>, this selector doesn't work and it's not obvious why without looking in browser inspector.

I just wrapped all rows in \<tbody\> so you can see it in editable html.