Update child.html
The example could be improved.  While fiddling with the 'border-top' field, changing it to 'border-left' or 'border-right' created a border that ran past ALL four link fields.  Changing it to 'border-bottom' created lines under the first and fourth link fields.  Changing 'border-top: 5px solid red;' to a simple 'color: red;' resulted in ALL four links turning red.  This is because the '<ol>' section is within the second '<li>' field, which itself is a direct child of '<ul>'.  

I think a better example would not have the '<ol>' section nested within the second link of the '<ul>' section, and use a simple 'color: red;' styling.  Then, removing the '>' demonstrates direct children vs all children.  And fiddling is simplified.